[08/10/2017] cleaning and rebuilding the routes
